+++ lib/metadata/metadata.c
@@ -1075,7 +1075,8 @@ bad:
 	return NULL;
 }
 
-uint64_t extents_from_size(struct cmd_context *cmd, uint64_t size,
+/* Rounds up by default */
+uint32_t extents_from_size(struct cmd_context *cmd, uint64_t size,
 			   uint32_t extent_size)
 {
 	if (size % extent_size) {
@@ -1094,7 +1095,7 @@ uint64_t extents_from_size(struct cmd_context *cmd, uint64_t size,
 		return 0;
 	}
 
-	return size / extent_size;
+	return (uint32_t) (size / extent_size);
 }
 
 /*
